Move+rename FreshQueue rpc strategies
Conflicts:
	Source/EasyNetQ/Rpc/FreshQueue/FreshQueueRpcFactory.cs
